Reworked tests on compilers to separate version detection
Tests on version detection are now separate from other
tests done on compiler.
